mainolimpiastart/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
django-bootstrap-modal-forms==3.0.4
django-bootstrap-v5==1.0.11
django-forms-bootstrap==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
filelock==3.11.0
idna==3.4
oauthlib==3.2.2
platformdirs==3.2.0
pycparser==2.21
PyJWT==2.7.0
pytils==0.4.1
pytz==2023.3
pywatchman==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.3
urllib3==2.0.2
utils==1.0.1
